fixed GetTotalNumberOfInputChannels() on laptop
for some unknown reasons the GetTotalNumberOfInputChannels() method on my laptop was returning a different value from instead of number of input buses declared in the constructor. This problem made every vector and array to go out of range. The problem has been solved by replacing that method with getBusesLayout().inputBuses.size()