reverb sends module working and changed variable "bus" to "channel"
The reverb sends module allows to send each channel after the processing but before the panning to an external reverb in the DAW.
Main thing to remember is that, if the VST is made for N track the reverb sends of track 1 and 2 have to be sent to output N+1 and N+2. This is because the output 1 and 2 are the main stereo output where each tack is panned to.

Also the variable "bus" is replaced with "channel" because the VST loads every input channel on a single input bus made of multiple mono channel.